fix: raw option-value shapes — user snowflake strings, float64 numbers
discordgo decodes Option.Value into interface{} verbatim as Discord
delivers it: USER options arrive as the user's snowflake STRING,
INTEGER options as JSON numbers (float64 through encoding/json),
string options as string, bool options as bool. The port asserted
*discordgo.User / int64 / int — all of them always failed:

- /gulag + /gulag-release: .(*discordgo.User) on the user option
  made EVERY invocation die with 'Please provide a valid user'
  (proven live at cutover). Now: the snowflake string, with the
  mention rendered <@ID> (v0.29.0 has no Mention helper).

- /gulag length option: .(int64) on the JSON number always missed
  and silently fell through to the 300s default. Now: float64 ->
  int64 with the unchanged (0, 10080] minutes clamp.

- cull 'days' option: .(int) always missed -> the option was
  silently the default no matter what the user passed. Now
  float64 -> int.

Test-pinned in TestHandleGulagOptionValueShapes (snowflake string +
float64 length reach the non-ephemeral with-reason success arm; a
non-string user value still yields 'Please provide a valid user').
Also routes invokerMember through the g.discord() seam (production
behavior unchanged — seam nil -> g.d) so the permVerify member fetch
is testable like its siblings. Parity checklist gains the
option-value-shape note alongside the Data-shape note.

+++ internal/handlers/cull/cull.go
@@ -349,8 +349,12 @@ func (h *Cull) HandleInteraction(i *discordgo.Interaction) Response {
 	// likewise the default).
 	days := DefaultDays
 	if o := commandOption(i, "days"); o != nil {
-		if v, ok := o.Value.(int); ok {
-			days = v
+		// INTEGER option: JSON numbers decode to float64 through
+		// interface{} — an .(int) assertion always failed and the
+		// option was (per Rust parity) silently treated as the
+		// default no matter what the user passed.
+		if v, ok := o.Value.(float64); ok {
+			days = int(v)
 		}
 	}
 	if !validateDays(days) {
